initial integration with multiple pattern matching
