+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Wed Oct 19 18:16:39 EDT 2022
+gradle.version=7.4.2
