+++ discord_bot/build.gradle
