+++ price_log/2022-10-19.txt
@@ -1,0 +1,201 @@
+2022-10-19T13:45:00.000Z Available funds: 100
+Buying: 143.105
+Current shares: 
+
+2022-10-19T13:46:00.000Z Available funds: 90
+Buying: 143.33
+Current shares: 143.105
+
+2022-10-19T13:48:00.000Z Available funds: 80
+Buying: 143.53
+Current shares: 143.105 143.33
+
+2022-10-19T13:52:00.000Z Available funds: 80.05101149505607
+Current shares: 143.105 143.33 143.53
+Sold 143.10
+
+2022-10-19T13:52:00.000Z Available funds: 90.07226140796627
+Current shares: 143.33 143.53
+Sold 143.53
+
+2022-10-19T13:56:00.000Z Available funds: 100.08098254101587
+Current shares: 143.33
+Sold 143.33
+
+2022-10-19T14:01:00.000Z Available funds: 100.08098254101587
+Buying: 143.55
+Current shares: 
+
+2022-10-19T14:05:00.000Z Available funds: 90.08098254101587
+Buying: 144.005
+Current shares: 143.55
+
+2022-10-19T14:06:00.000Z Available funds: 80.08098254101587
+Buying: 144.15
+Current shares: 143.55 144.005
+
+2022-10-19T14:07:00.000Z Available funds: 70.08098254101587
+Buying: 144.345
+Current shares: 143.55 144.005 144.15
+
+2022-10-19T14:12:00.000Z Available funds: 60.08098254101587
+Buying: 144.215
+Current shares: 143.55 144.005 144.15 144.345
+
+2022-10-19T14:14:00.000Z Available funds: 50.08098254101587
+Buying: 144.57
+Current shares: 143.55 144.005 144.15 144.345 144.215
+
+2022-10-19T14:15:00.000Z Available funds: 40.08098254101587
+Buying: 144.72
+Current shares: 143.55 144.005 144.15 144.345 144.215 144.57
+
+2022-10-19T14:16:00.000Z Available funds: 40.134274077066024
+Current shares: 143.55 144.005 144.15 144.345 144.215 144.57 144.72
+Sold 143.55
+
+2022-10-19T14:16:00.000Z Available funds: 50.14572048705562
+Current shares: 144.005 144.15 144.345 144.215 144.57 144.72
+Sold 144.15
+
+2022-10-19T14:16:00.000Z Available funds: 60.152654578516284
+Current shares: 144.005 144.345 144.215 144.57 144.72
+Sold 144.22
+
+2022-10-19T14:17:00.000Z Available funds: 70.17244555799616
+Current shares: 144.005 144.345 144.57 144.72
+Sold 144.00
+
+2022-10-19T14:18:00.000Z Available funds: 70.17244555799616
+Buying: 144.43
+Current shares: 144.345 144.57 144.72
+
+2022-10-19T14:24:00.000Z Available funds: 60.172445557996156
+Buying: 144.91
+Current shares: 144.345 144.57 144.72 144.43
+
+2022-10-19T14:28:00.000Z Available funds: 50.172445557996156
+Buying: 144.97
+Current shares: 144.345 144.57 144.72 144.43 144.91
+
+2022-10-19T14:31:00.000Z Available funds: 50.19218992046108
+Current shares: 144.345 144.57 144.72 144.43 144.91 144.97
+Sold 144.34
+
+2022-10-19T14:31:00.000Z Available funds: 60.20603745906109
+Current shares: 144.57 144.72 144.43 144.91 144.97
+Sold 144.43
+
+2022-10-19T14:44:00.000Z Available funds: 60.20603745906109
+Buying: 143.995
+Current shares: 144.57 144.72 144.91 144.97
+
+2022-10-19T14:46:00.000Z Available funds: 50.20603745906109
+Buying: 144.075
+Current shares: 144.57 144.72 144.91 144.97 143.995
+
+2022-10-19T14:47:00.000Z Available funds: 40.20603745906109
+Buying: 144.13
+Current shares: 144.57 144.72 144.91 144.97 143.995 144.075
+
+2022-10-19T15:01:00.000Z Available funds: 30.206037459061093
+Buying: 143.735
+Current shares: 144.57 144.72 144.91 144.97 143.995 144.075 144.13
+
+2022-10-19T15:05:00.000Z Available funds: 20.206037459061093
+Buying: 143.525
+Current shares: 144.57 144.72 144.91 144.97 143.995 144.075 144.13 143.735
+
+2022-10-19T15:07:00.000Z Available funds: 10.206037459061093
+Buying: 143.695
+Current shares: 144.57 144.72 144.91 144.97 143.995 144.075 144.13 143.735 143.525
+
+2022-10-19T15:12:00.000Z Available funds: 10.216125468244659
+Current shares: 144.57 144.72 144.91 144.97 143.995 144.075 144.13 143.735 143.525 143.695
+Sold 143.74
+
+2022-10-19T15:12:00.000Z Available funds: 20.2289999593543
+Current shares: 144.57 144.72 144.91 144.97 143.995 144.075 144.13 143.525 143.695
+Sold 143.69
+
+2022-10-19T15:19:00.000Z Available funds: 20.2289999593543
+Buying: 144.05
+Current shares: 144.57 144.72 144.91 144.97 143.995 144.075 144.13 143.525
+
+2022-10-19T15:20:00.000Z Available funds: 20.247463641639616
+Current shares: 144.57 144.72 144.91 144.97 143.995 144.075 144.13 143.525 144.05
+Sold 143.53
+
+2022-10-19T15:22:00.000Z Available funds: 20.247463641639616
+Buying: 143.92
+Current shares: 144.57 144.72 144.91 144.97 143.995 144.075 144.13 144.05
+
+2022-10-19T15:39:00.000Z Available funds: 10.247463641639616
+Buying: 143.87
+Current shares: 144.57 144.72 144.91 144.97 143.995 144.075 144.13 144.05 143.92
+
+2022-10-19T15:53:00.000Z Available funds: 10.261353012798335
+Current shares: 144.57 144.72 144.91 144.97 143.995 144.075 144.13 144.05 143.92 143.87
+Sold 144.00
+
+2022-10-19T15:53:00.000Z Available funds: 20.265862830324178
+Current shares: 144.57 144.72 144.91 144.97 144.075 144.13 144.05 143.92 143.87
+Sold 144.13
+
+2022-10-19T15:53:00.000Z Available funds: 30.28497066801178
+Current shares: 144.57 144.72 144.91 144.97 144.075 144.05 143.92 143.87
+Sold 143.92
+
+2022-10-19T15:56:00.000Z Available funds: 40.29052333155508
+Current shares: 144.57 144.72 144.91 144.97 144.075 144.05 143.87
+Sold 144.07
+
+2022-10-19T15:56:00.000Z Available funds: 50.310332881843536
+Current shares: 144.57 144.72 144.91 144.97 144.05 143.87
+Sold 143.87
+
+2022-10-19T16:30:00.000Z Available funds: 50.310332881843536
+Buying: 143.49
+Current shares: 144.57 144.72 144.91 144.97 144.05
+
+2022-10-19T16:31:00.000Z Available funds: 40.310332881843536
+Buying: 143.645
+Current shares: 144.57 144.72 144.91 144.97 144.05 143.49
+
+2022-10-19T16:32:00.000Z Available funds: 30.310332881843536
+Buying: 143.725
+Current shares: 144.57 144.72 144.91 144.97 144.05 143.49 143.645
+
+2022-10-19T16:33:00.000Z Available funds: 20.310332881843536
+Buying: 143.77
+Current shares: 144.57 144.72 144.91 144.97 144.05 143.49 143.645 143.725
+
+2022-10-19T16:34:00.000Z Available funds: 10.310332881843536
+Buying: 143.81
+Current shares: 144.57 144.72 144.91 144.97 144.05 143.49 143.645 143.725 143.77
+
+2022-10-19T16:36:00.000Z Available funds: 10.328452611441417
+Current shares: 144.57 144.72 144.91 144.97 144.05 143.49 143.645 143.725 143.77 143.81
+Sold 143.49
+
+2022-10-19T16:36:00.000Z Available funds: 20.33019204438628
+Current shares: 144.57 144.72 144.91 144.97 144.05 143.645 143.725 143.77 143.81
+Sold 143.72
+
+2022-10-19T16:38:00.000Z Available funds: 20.33019204438628
+Buying: 143.6
+Current shares: 144.57 144.72 144.91 144.97 144.05 143.645 143.77 143.81
+
+2022-10-19T16:51:00.000Z Available funds: 10.33019204438628
+Buying: 143.14
+Current shares: 144.57 144.72 144.91 144.97 144.05 143.645 143.77 143.81 143.6
+
+2022-10-19T18:31:00.000Z Available funds: 10.340671295469136
+Current shares: 144.57 144.72 144.91 144.97 144.05 143.645 143.77 143.81 143.6 143.14
+Sold 143.14
+
+2022-10-19T18:34:00.000Z Available funds: 10.340671295469136
+Buying: 143.49
+Current shares: 144.57 144.72 144.91 144.97 144.05 143.645 143.77 143.81 143.6
+
+End of day fund: 0.34
